Search now works with the new location format. Search is also no longer case sensitive.
When viewing a post, the user can see the address of the sale. The user can also click on the address which will open their google maps app with the location (some addresses wont allow for the button to be pressed if its funky, but most work just fine).

Also found a new bug. Once a post is deleted, View my posts will be blank until the user logs out and logs back in. Not sure whats causing this